server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2


pymupdf==1.24.9
python-docx==1.1.2
pytesseract==0.3.13
Pillow==10.4.0
unstructured==0.15.12